added detail for main screen & some other details

+++ SpendTrack/Model/SingleResult.swift
@@ -10,10 +10,9 @@ import Foundation
 struct SingleResult: Decodable{
     var title: String
     var link: String
-    var price: Int
+    var price: Double
     var price_raw: String
     var snippet: String?
-    //var price_parsed: Price
     var merchant: String
     var image: String
 }
